ultimos cambios con excepciones

+++ RentCar/Modelo/Arrendador.java
@@ -4,7 +4,7 @@ import java.util.ArrayList;
 
 public class Arrendador extends Usuario {
 
-    public Arrendador(String nombre, String apellido, int edad, String direccion, long telefono, String correoElectronico, String tipoIdentificacion, int numeroDelIdentificacion) {
+    public Arrendador(String nombre, String apellido, int edad, String direccion, long telefono, String correoElectronico, String tipoIdentificacion, Long numeroDelIdentificacion) {
         super(nombre, apellido, edad, direccion, telefono, correoElectronico, tipoIdentificacion, numeroDelIdentificacion);
 
     }
