+++ RentCar/Modelo/Arrendatario.java
@@ -17,12 +17,12 @@ public class Arrendatario extends Usuario{
 //        this.preferencial=preferencial;
 //        this.contrato=contrato;
 //    }
-public Arrendatario(String nombre, String apellido, int edad, String direccion, long telefono, String correoElectronico, String tipoIdentificacion, int numeroDelIdentificacion,  boolean licenciaConduccion, boolean preferencial) {
+public Arrendatario(String nombre, String apellido, int edad, String direccion, long telefono, String correoElectronico, String tipoIdentificacion, Long numeroDelIdentificacion,  boolean licenciaConduccion, boolean preferencial) {
     super(nombre, apellido, edad, direccion, telefono, correoElectronico, tipoIdentificacion, numeroDelIdentificacion);
     this.licenciaConduccion= licenciaConduccion;
     this.preferencial=preferencial;
 }
-    public Arrendatario(String nombre, String apellido, int edad, String direccion, long telefono, String correoElectronico, String tipoIdentificacion, int numeroDelIdentificacion,  boolean licenciaConduccion, boolean preferencial, ArrayList<Vehiculo> vehiculos) {
+    public Arrendatario(String nombre, String apellido, int edad, String direccion, long telefono, String correoElectronico, String tipoIdentificacion, Long numeroDelIdentificacion,  boolean licenciaConduccion, boolean preferencial, ArrayList<Vehiculo> vehiculos) {
         super(nombre, apellido, edad, direccion, telefono, correoElectronico, tipoIdentificacion, numeroDelIdentificacion, vehiculos);
         this.licenciaConduccion= licenciaConduccion;
         this.preferencial=preferencial;
